Plugins read Boolean and Integer parameters as int
OpenFX passes Boolean and Integer parameter values as int, and the host
now writes them that way. The plugins were receiving them into bool and
int64_t variables, which the old host had matched by writing whatever
type the JSON held. Written as int, a bool got four bytes into one, and
on Windows that took the gradient generator's stack with it: the test
render hung at the first gradient frame.

+++ plugins/DrawPlugin.cpp
@@ -188,8 +188,8 @@ OfxStatus BoxPlugin::_render(
     const OfxRectI& renderWindow,
     OfxPropertySetHandle inArgs)
 {
-    int64_t pos1[2] = { 0, 0 };
-    int64_t pos2[2] = { 0, 0 };
+    int pos1[2] = { 0, 0 };
+    int pos2[2] = { 0, 0 };
     double color[4] = { 0.0, 0.0, 0.0, 0.0 };
     int fill = 0;
     _paramSuite->paramGetValue(_param(handle, "pos1"), &pos1[0], &pos1[1]);
@@ -282,8 +282,8 @@ OfxStatus LinePlugin::_render(
     const OfxRectI& renderWindow,
     OfxPropertySetHandle inArgs)
 {
-    int64_t pos1[2] = { 0, 0 };
-    int64_t pos2[2] = { 0, 0 };
+    int pos1[2] = { 0, 0 };
+    int pos2[2] = { 0, 0 };
     double color[4] = { 0.0, 0.0, 0.0, 0.0 };
     int skipFirstPoint = 0;
     _paramSuite->paramGetValue(_param(handle, "pos1"), &pos1[0], &pos1[1]);
@@ -379,9 +379,9 @@ OfxStatus TextPlugin::_render(
     const OfxRectI& renderWindow,
     OfxPropertySetHandle inArgs)
 {
-    int64_t pos[2] = { 0, 0 };
+    int pos[2] = { 0, 0 };
     std::string text;
-    int64_t fontSize = 16;
+    int fontSize = 16;
     std::string fontName;
     double color[4] = { 0.0, 0.0, 0.0, 0.0 };
     _paramSuite->paramGetValue(_param(handle, "pos"), &pos[0], &pos[1]);
